Refactor config.Extensions to its own file
