+++ config/config.go
@@ -8,6 +8,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/keep-network/keep-common/pkg/chain/ethereum"
 	"github.com/keep-network/keep-core/pkg/net/libp2p"
+	eth "github.com/keep-network/keep-ecdsa/pkg/chain"
 	"github.com/keep-network/keep-ecdsa/pkg/client"
 	"github.com/keep-network/keep-ecdsa/pkg/ecdsa/tss"
 )
@@ -28,7 +29,7 @@ type Config struct {
 	TSS                    tss.Config
 	Metrics                Metrics
 	Diagnostics            Diagnostics
-	Extensions             Extensions
+	Extensions             eth.Config
 }
 
 // SanctionedApplications contains addresses of applications approved by the
@@ -71,24 +72,6 @@ type Metrics struct {
 // Diagnostics stores diagnostics-related configuration.
 type Diagnostics struct {
 	Port int
-}
-
-// Extensions stores app-specific extensions configuration.
-type Extensions struct {
-	TBTC TBTC
-}
-
-// TBTC stores configuration of application extension responsible for
-// executing signer actions specific for TBTC application.
-type TBTC struct {
-	TBTCSystem string
-	BTCRefunds BTCRefunds
-}
-
-// BTCRefunds stores configuration related to recovering BTC from a closed keep.
-type BTCRefunds struct {
-	BeneficiaryAddress string
-	MaxFeePerVByte     int32
 }
 
 // ReadConfig reads in the configuration file in .toml format. Ethereum key file
